refactor database config into its own class
